requirements.txt
appdirs==1.4.3
certifi==2019.3.9
chardet==3.0.4
idna==2.8
numpy==1.16.2
pandas==0.24.2
praw==6.1.1
prawcore==1.0.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
update-checker==0.16
urllib3==1.24.2
websocket-client==0.56.0
